python/requirements.txt
# requirements for odoo-arduino module
pyserial==3.5

